+++ pyproject.toml
@@ -1,0 +1,5 @@
+[tool.pytest.ini_options]
+addopts = [
+    "--import-mode=importlib",
+]
+pythonpath = "src"
