+++ test/test_download.py
@@ -1,9 +1,8 @@
 import os
 import moto
 import boto3
-import pyfakefs
 import pytest
-from download import get_bucket, download_shoot_folder
+from transferrer.download import get_bucket, download_shoot_folder
 
 HERE = os.path.dirname(os.path.abspath(__file__))
